Backend/requirements.txt
uvicorn==0.34.0
mysql-connector-python==9.2.0
fastapi==0.115.12
pydantic==2.11.1